Trying to implement message callback (#165).
