+++ playpen/playpen.py
@@ -689,6 +689,7 @@ webview_settings = webview.get_settings()
 webview_settings.set_enable_javascript(True)
 webview_settings.set_enable_developer_extras(True)
 webview_settings.set_allow_universal_access_from_file_urls(True)
+webview_settings.set_allow_file_access_from_file_urls(True)
 webview_settings.set_enable_webgl(True)
 webview_settings.set_enable_webaudio(True)
 webview.load_uri("http://csound.com")
